【界面】 eos内存柱状图

+++ app/src/main/java/io/goldstone/blockchain/kernel/network/goldstoneapi.kt
@@ -574,9 +574,9 @@ object GoldStoneAPI {
 	
 	fun getEOSRAMTradeData(
 		errorCallback: (Exception) -> Unit,
-		hold: (ArrayList<Double>) -> Unit
+		hold: (ArrayList<Float>) -> Unit
 	) {
-		requestData<Double>(
+		requestData<Float>(
 			APIPath.getEOSRAMTradeData(APIPath.currentUrl),
 			"data",
 			errorCallback = errorCallback,
